+++ sdk-java/src/main/java/ly/count/sdk/java/internal/UserEditorImpl.java
@@ -279,7 +279,6 @@ public class UserEditorImpl implements UserEditor {
         } catch (JSONException e) {
             L.e("[UserEditorImpl] Exception while committing changes to User profile" + e);
         }
-
         return user;
     }
 }