+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 
 * Added the user profiles feature interface, and it is accessible through "Countly::instance()::userProfile()" call.
 
+* Fixed a bug where setting custom user properties would not work.
+
+* Deprecated "Countly::backendMode()" call, use "Countly::backendM" instead via "instance()" call.
 * The following methods are deprecated from the "UserEditor" interface:
   * "commit()" instead use "Countly::userProfile::save" via "instance()" call
   * "pushUnique(String, Object)" instead use "Countly::userProfile::pushUnique" via "instance()" call